Add Node class and unit tests
This is a wrapper class that will allow string values to be passed by
reference.

Will make it easier to check winning patterns in the Connect Four game,
since nodes are the same, whether they are in a row, column, or
diagonal.  For example one node on a row will also be present on a
column and two diagonals (uphill and downhill).  If the node's value
changes in one, it needs to change on all.