Add Grid class and unit tests
Is a 2-dimensional array, but with row, column, and diagonal views.  The
different views will be needed to check for the 4 in a row winning
condition.  Every possible line (up, down, and diagonals) is represented
as an array (stored in all_majors), so will just need to loop through
that to see if any 4 in a row is present.